requirements.txt
requests
pandas
pyarrow
netCDF4
python-dotenv
art
xarray
